# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Shared Music API",
    "version": "1.0.0",
    "description": "API to share music with friends using Spotify."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/oAuthLogin": {
      "get": {
        "summary": "LOGIN with Spotify, using OAuth 2.0 protocol.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Redirects the user to the Spotify login page for authentication.",
        "tags": [
          "Spotify OAuth authentication"
        ],
        "responses": {
          "302": {
            "description": "Redirect to Spotify authentication page.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "https://accounts.spotify.com/authorize"
              }
            }
          }
        }
      }
    },
    "/callback": {
      "get": {
        "summary": "Manage the redirection after Spotify authentication.",
        "description": "Handles the callback from Spotify OAuth flow, exchanges code for access token.",
        "tags": [
          "Spotify OAuth authentication"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "description": "The authorization code returned by Spotify.",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "description": "The state parameter sent by the client to avoid CSRF. In our case, we use the state to pass the username of the user who initiated the OAuth flow.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the access token.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Access Token: your_access_token_here"
                }
              }
            }
          },
          "500": {
            "description": "Error while getting the Spotify token."
          }
        }
      }
    },
    "/getUserPersonality": {
      "get": {
        "summary": "Get user's music listening personality.",
        "description": "Analyzes the user's top listened tracks to deduce their musical personality based on various audio features.",
        "tags": [
          "Spotify features"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the user's music listening personality.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "dance": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 10,
                      "description": "Danceability score on a scale of 1 to 10."
                    },
                    "agitation": {
                      "type": "integer",
                      "description": "Average tempo of top tracks."
                    },
                    "preference": {
                      "type": "string",
                      "description": "User's preference for instrumental or vocal tracks."
                    },
                    "attitude": {
                      "type": "string",
                      "description": "User's overall positivity or negativity in music choice."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No liked songs, personality impossible to define."
          },
          "500": {
            "description": "Error while getting the Spotify songs or audio statistics."
          }
        }
      }
    },
    "/syncCurrentTrack": {
      "get": {
        "summary": "Synchronize the track currently played with all the members of the user's group.",
        "description": "Get the user's current track and its progress, then sync this track with all the members of the user's group. Requires the user to be authenticated and a member of a group.",
        "tags": [
          "Spotify features"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the result of the API calls for the track synchronization with the group members.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "user": {
                        "type": "string",
                        "description": "Username of the user for which the API call was made."
                      },
                      "success": {
                        "type": "boolean",
                        "description": "true if the API call was successful, false otherwise."
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The user is not part of any group."
          },
          "500": {
            "description": "Error while syncing the current track with the group members."
          }
        }
      }
    },
    "/createPlaylistFromUserTopTracks": {
      "get": {
        "summary": "Create a playlist from another user's top tracks.",
        "description": "Get the top 10 tracks of a specified Spotify user and create a new playlist in the current user's Spotify account containing these tracks.",
        "tags": [
          "Spotify features"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "required": true,
            "description": "Username of the Spotify user from which to get the top tracks.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tracks added successfully to the playlist!"
          },
          "400": {
            "description": "User or spotify token not found."
          },
          "500": {
            "description": "Error while getting the user's top tracks or creating a playlist from them."
          }
        }
      }
    }
  },
...
}